Move equality members to separate file
